Just move to Historic, don't try to also obsolete
The IESG consensus seems to be that 4757 should move to Historic,
even though it's a bit weird to do so (since it's just an
Informational document describing the microsoft implementation ...
that everyone else then went and implemented so that they could
interoperate).  That means there's no need to also obsolete it,
and it's worth noting that we know we're making the status change
even though there is still some usage.